requirements.txt
Django>=3.1,<3.2
djangorestframework==3.12.2
whitenoise==4.1.3
uwsgi==2.0.19.1
graphene-django==2.15.0
django-cors-headers==3.7.0
psycopg2==2.8.5
dj-database-url==0.5.0
google-api-python-client
oauth2client
matplotlib
django-filter
graphene-django-extras